[WebGPU plugin EP packaging] Remove explicit ORT package dependency (#28384)
### Description
<!-- Describe your changes. -->

This pull request refactors how the minimum required ONNX Runtime (ORT)
version is handled and communicated for the WebGPU plugin EP packages
(both Python and C#). Instead of declaring a hard dependency on a
specific ORT version in package manifests, the minimum compatible
version is now injected into package READMEs at build time, and ORT
version compatibility will be checked at runtime. The packaging scripts
are updated to use a shared template utility, and the CI/test setup is
adjusted accordingly.

The most important changes are:

**Minimum ORT Version Handling and Documentation:**
* The Python and C# plugin EP packages no longer declare a hard
dependency on a specific ONNX Runtime package version in their manifests
(`pyproject.toml`, `.csproj`). Instead, the minimum required ORT version
is injected into the package `README.md` during packaging, and users are
instructed to install or reference the correct ORT version themselves.

**Packaging and Build Script Refactoring:**
* A new shared utility, `_packaging_utils.py`, is added for
template-based file generation, and both Python (`build_wheel.py`) and
C# (`pack_nuget.py`) packaging scripts are updated to use this helper
for injecting the minimum ORT version into `README.md`. Duplicate
template code is removed.
* The C# packaging script and project file are refactored to remove the
mechanism for passing the minimum ORT version via MSBuild properties,
since the dependency is now documented rather than enforced at
build/package time.

**Test and CI Adjustments:**
* The Python test pipelines for Linux, macOS, and Windows are updated to
explicitly install `onnxruntime` before testing the plugin EP, since it
is no longer a transitive dependency.

These changes improve clarity for users about required dependencies,
centralize version management, and simplify the packaging scripts.

NOTE: The WebGPU EP ORT version compatibility check does not actually
use the value in `./plugin-ep-webgpu/MIN_ONNXRUNTIME_VERSION` yet. That
can be done in another PR.

### Motivation and Context
<!-- - Why is this change required? What problem does it solve?
- If it fixes an open issue, please link to the issue here. -->

There are multiple packages that can provide an ORT library. Even though
the basic, CPU EP-only ORT package would typically be sufficient, we
don't want to mandate usage of a specific package flavor.
