+++ plugin-ep-webgpu/python/pyproject.toml.in
@@ -9,9 +9,6 @@ description = "ONNX Runtime WebGPU Plugin Execution Provider"
 readme = "onnxruntime_ep_webgpu/README.md"
 license = {text = "MIT"}
 requires-python = ">=3.11"
-dependencies = [
-  "onnxruntime>=@min_onnxruntime_version@",
-]
 
 [tool.setuptools.packages.find]
 include = ["onnxruntime_ep_webgpu*"]
